requirements.txt
python-telegram-bot==13.0
pymongo==3.11.0
requests==2.24.0
python-dotenv==0.15.0
gunicorn==20.0.4
dnspython==1.16.0
uuid==1.30
py7zr==0.10.1
